friender-backend/requirements.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
bcrypt==3.2.2
blinker==1.4
boto3==1.24.38
botocore==1.27.38
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.1
executing==0.8.3
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
forex-python==1.8
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.0
s3transfer==0.6.0
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.37
stack-data==0.3.0
traitlets==5.3.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
